Rename Variables, remove unnecessary imports, delete deprecated files
