+++ player.tscn
@@ -1,10 +1,40 @@
-[gd_scene load_steps=3 format=3 uid="uid://dgcmv0e74kcr6"]
+[gd_scene load_steps=3 format=3 uid="uid://i5vtdujt6dow"]
 
-[ext_resource type="Script" path="res://Player.cs" id="1_ol0sv"]
 [ext_resource type="Texture2D" uid="uid://dsje2yrpefe4h" path="res://icon.svg" id="2_wwi64"]
 
+[sub_resource type="CSharpScript" id="CSharpScript_7ebf1"]
+script/source = "using Godot;
+using System;
+
+public partial class Player : CharacterBody2D
+{
+	public const float Speed = 300.0f;
+	public const float JumpVelocity = -400.0f;
+
+	public override void _PhysicsProcess(double delta)
+	{
+		Vector2 velocity = Velocity;
+
+		// Get the input direction and handle the movement/deceleration.
+		// As good practice, you should replace UI actions with custom gameplay actions.
+		Vector2 direction = Input.GetVector(\"left\", \"right\", \"up\", \"down\");
+		if (direction != Vector2.Zero)
+		{
+			velocity = direction * Speed;
+		}
+		else
+		{
+			velocity.X = Mathf.MoveToward(Velocity.X, 0, Speed);
+			velocity.Y = Mathf.MoveToward(Velocity.Y, 0, Speed);
+		}
+
+		Velocity = velocity;
+		MoveAndSlide();
+	}
+}"
+
 [node name="Player" type="CharacterBody2D"]
-script = ExtResource("1_ol0sv")
+script = SubResource("CSharpScript_7ebf1")
 
 [node name="Icon" type="Sprite2D" parent="."]
 texture = ExtResource("2_wwi64")
