+++ project.godot
@@ -11,7 +11,7 @@ config_version=5
 [application]
 
 config/name="untitled-plant-game"
-run/main_scene="res://Tutorial_Player/Tutorial.tscn"
+run/main_scene="res://Player/Tutorial.tscn"
 config/features=PackedStringArray("4.2", "C#", "GL Compatibility")
 config/icon="res://icon.svg"
